@@ -2,6 +2,7 @@
   "name": "booking-for-pets",
   "jest": {
     "testEnvironment": "jsdom",
+    "globalSetup": "./configs/jest.setup.ts",
     "moduleFileExtensions": [
       "ts",
       "tsx",
@@ -16,7 +17,11 @@
     ],
     "moduleNameMapper": {
       "^.+\\.(css|less|scss)$": "babel-jest"
-    }
+    },
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-dom",
+      "./configs/jest.setup.ts"
+    ]
   },
   "version": "0.1.0",
   "description": "A React.js project using Typescript, with which you can book a hotel room for your pet",
@@ -38,12 +43,11 @@
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.7",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18.0.5",
     "@typescript-eslint/eslint-plugin": "^5.8.1",
     "@typescript-eslint/parser": "^5.8.1",
     "babel-jest": "^28.1.0",
@@ -61,8 +65,7 @@
     "jest": "^27.4.5",
     "json-server": "^0.17.0",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-hot-loader": "^4.13.0",
     "rimraf": "^3.0.2",
     "sass-loader": "^12.4.0",
@@ -77,20 +80,28 @@
     "@emotion/react": "^11.8.1",
     "@emotion/styled": "^11.8.1",
     "@fontsource/roboto": "^4.5.3",
+    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.5.1",
     "@mui/lab": "^5.0.0-alpha.73",
     "@mui/material": "^5.8.1",
     "@reduxjs/toolkit": "^1.8.2",
+    "@types/debounce": "^1.2.1",
     "@types/json-server": "^0.14.4",
-    "@types/react-redux": "^7.1.23",
+    "@types/react": "^18.0.12",
+    "@types/react-redux": "^7.1.24",
     "@types/react-router-redux": "^5.0.21",
     "axios": "^0.26.1",
+    "crypto-browserify": "^3.12.0",
     "crypto-js": "^4.1.1",
     "date-fns": "^2.28.0",
+    "debounce": "^1.2.1",
     "formik": "^2.2.9",
     "history": "^5.3.0",
     "json-server-auth": "^2.1.0",
+    "njwt": "^1.2.0",
+    "node-polyfill-webpack-plugin": "^1.1.4",
     "notistack": "^1.0.10",
+    "react": "^18.1.0",
     "react-copy-to-clipboard": "^5.0.4",
     "react-dropzone": "^12.0.4",
     "react-redux": "^7.2.6",
@@ -103,6 +114,7 @@
     "redux-thunk": "^2.4.1",
     "sass": "^1.49.9",
     "ts-node": "^10.7.0",
+    "whatwg-fetch": "^3.6.2",
     "yup": "^0.32.11"
   }
 }
