@@ -2,10 +2,6 @@ import React from "react";
 import { LinearProgress, Box, Alert, AlertTitle } from "@mui/material";
 import { preloaderStyles } from "./auxiliaryStyles";
 import { IPreloader } from "../../server/api/api-models";
-import {
-  RESOLVED_PROMISE_STATUS,
-  REJECTED_PROMISE_STATUS,
-} from "../../helpers/consts";
 
 export const Loader = () => {
   return (
@@ -16,34 +12,29 @@ export const Loader = () => {
 };
 
 export const Preloader = ({
-  promiseName,
-  promiseState,
-  sub,
+  isLoading,
+  error,
   modal,
+  children,
 }: IPreloader) => {
   return (
     <>
-      {!promiseState[promiseName] ||
-      promiseState[promiseName]?.status === RESOLVED_PROMISE_STATUS ? (
-        sub
-      ) : promiseState[promiseName]?.status === REJECTED_PROMISE_STATUS ? (
-        promiseState[promiseName]?.error?.message ? (
-          <>
-            <Alert
-              severity="error"
-              style={modal ? preloaderStyles.modal : preloaderStyles.main}
-            >
-              <AlertTitle>Error</AlertTitle>
-              {promiseState[promiseName]?.error?.message}
-            </Alert>
-            <br />
-            {sub}
-          </>
-        ) : (
-          <>{sub}</>
-        )
-      ) : (
+      {isLoading ? (
         <Loader />
+      ) : error ? (
+        <>
+          <Alert
+            severity="error"
+            style={modal ? preloaderStyles.modal : preloaderStyles.main}
+          >
+            <AlertTitle>Error</AlertTitle>
+            {error}
+          </Alert>
+          <br />
+          {children}
+        </>
+      ) : (
+        children
       )}
     </>
   );
