@@ -8,152 +8,155 @@ import {
 } from "@mui/material";
 import { CardActionArea, IconButton } from "@mui/material";
 import { Link } from "react-router-dom";
-import { RootState } from "../../helpers/types";
-import { useSelector } from "react-redux";
 import { hotelCardStyles } from "./hotelsStyles";
 import {
-  actionFullHotelUpdate as hotelUpdate,
-  actionUpdateWishlists as onUnsave,
-} from "./../../actions/thunks";
-import { truncText } from "../../helpers/functions";
-import { links, sendSnackBarMessages } from "../../helpers/consts";
+  formateUser,
+  truncText,
+  unsaveHotelFromWishlist,
+} from "../../helpers/functions";
+import { links, siteCurrencyList } from "../../helpers/consts";
 import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 import FavoriteIcon from "@mui/icons-material/Favorite";
-import ModalWindow from "./../Auxiliary/ModalWindow";
+import { ModalWindow } from "./../Auxiliary/ModalWindow";
 import { SelectWishlist } from "./../Wishlist/SelectWishlist";
 import useSnackBar from "./../Auxiliary/SnackBar";
-import { CurrencyModel, WishlistModel } from "src/server/api/api-models";
+import { CurrencyModel, WishlistModel } from "../../server/api/api-models";
+import { usersAPI } from "../../store/reducers/UserService";
+import { currencyAPI } from "../../store/reducers/CurrencyService";
+import { hotelAPI } from "../../store/reducers/HotelService";
+import { Preloader } from "../Auxiliary/Preloader";
 
 export const HotelCard = ({ hotelData }) => {
-  const auth = useSelector((state: RootState) => state.promise);
-  const currencyList = useSelector((state: RootState) => state.promise);
+  const currentUser = formateUser();
 
   const [openWishlistsWindow, setOpenWishlistsWindow] = useState(false);
   const [, sendSnackbar] = useSnackBar();
 
-  const currentCurrency = (currencyList?.currency || []).find(
-    (currency: CurrencyModel) => auth?.payload?.currencyId === currency?.id
+  const currentCurrency = (siteCurrencyList || []).find(
+    (currency: CurrencyModel) => currentUser?.currencyId === currency?.id
   );
 
   const updateWishlistWindow = (value: boolean) => {
     setOpenWishlistsWindow(value);
   };
 
-  const currentUserWishlists = auth?.payload?.wishlists;
+  const currentUserWishlists = currentUser?.wishlists;
 
   const isSaved = (currentUserWishlists || []).find((wishlist: WishlistModel) =>
     wishlist.hotelsId.includes(hotelData.id)
   );
 
-  function handleUnsave() {
-    const isSavedIndex = (currentUserWishlists || []).findIndex(
-      (wishlist: WishlistModel) => wishlist.hotelsId.includes(hotelData.id)
-    );
+  const [userUpdate, { isLoading: userUpdateLoading, error: userUpdateError }] =
+    usersAPI.useUpdateUserMutation();
 
-    const filteredHotelsId = isSaved.hotelsId.filter(
-      (hotelId: string) => hotelId !== hotelData.id
-    );
+  const {
+    data: currencyList,
+    error: currencyError,
+    isLoading: currencyLoading,
+  } = currencyAPI.useFetchAllCurrencyQuery("");
 
-    const filteredUserWishlists = [...currentUserWishlists];
-
-    filteredUserWishlists.splice(isSavedIndex, 1);
-
-    const filteredCurrentUserWishlist = {
-      ...currentUserWishlists[isSavedIndex],
-      hotelsId: filteredHotelsId,
-    };
-
-    filteredUserWishlists.push(filteredCurrentUserWishlist);
-
-    if (typeof sendSnackbar === "function") {
-      sendSnackbar({
-        msg: sendSnackBarMessages.removedFromWishlistMessage(isSaved?.name),
-        variant: "error",
-      });
-    }
-
-    onUnsave(filteredUserWishlists);
-  }
-
-  const [saved] = useState(isSaved ? true : false);
+  const [
+    updateHotel,
+    { isLoading: updateHotelLoading, error: updateHotelError },
+  ] = hotelAPI.useUpdateHotelMutation();
 
   return (
-    <Card sx={hotelCardStyles.main}>
-      {openWishlistsWindow && (
-        <ModalWindow
-          title={"Your choice"}
-          body={
-            <SelectWishlist
-              modalWindowState={updateWishlistWindow}
-              currentHotel={hotelData}
-            />
-          }
-          modalWindowState={updateWishlistWindow}
-        />
-      )}
-      <Link to={`/hotels/hotel/${hotelData.id}`}>
-        <CardActionArea onClick={() => hotelUpdate({ id: hotelData.id })}>
-          <CardMedia
-            component="img"
-            height="140"
-            image={hotelData.image || links.noImage}
-            alt={`Hotel ${hotelData.index + 1}`}
+    <Preloader
+      isLoading={userUpdateLoading || currencyLoading || updateHotelLoading}
+      error={
+        userUpdateError?.data || currencyError?.data || updateHotelError?.data
+      }
+    >
+      <Card sx={hotelCardStyles.main}>
+        {openWishlistsWindow && (
+          <ModalWindow
+            title={"Your choice"}
+            body={
+              <SelectWishlist
+                modalWindowState={updateWishlistWindow}
+                currentHotel={hotelData}
+              />
+            }
+            modalWindowState={updateWishlistWindow}
           />
-          <CardContent sx={{ height: "20vh" }}>
-            <Typography gutterBottom variant="h6" component="div">
-              {hotelData.title}
-            </Typography>
-            <Typography variant="body1" color="text.secondary">
-              {truncText(hotelData.description, 70)}
-            </Typography>
-          </CardContent>
-        </CardActionArea>
-      </Link>
-      <CardActions>
-        {saved ? (
-          <>
-            <Typography variant="body1" color="text.secondary">
-              <b>
-                {currentCurrency?.sign}
-                {hotelData.price *
-                  currencyList?.exchangeList[currentCurrency?.name]?.toFixed(
-                    1
-                  )}{" "}
-                / day
-              </b>
-            </Typography>
-            <IconButton
-              color="secondary"
-              sx={hotelCardStyles.rightAlign}
-              disabled={!auth?.token}
-              onClick={handleUnsave}
-            >
-              <FavoriteIcon />
-            </IconButton>
-          </>
-        ) : (
-          <>
-            <Typography variant="body1" color="text.secondary">
-              <b>
-                {currentCurrency?.sign || "$"}
-                {hotelData.price *
-                  currencyList?.exchangeList[currentCurrency?.name]?.toFixed(
-                    1
-                  ) || hotelData.price}{" "}
-                / day
-              </b>
-            </Typography>
-            <IconButton
-              color="secondary"
-              sx={hotelCardStyles.rightAlign}
-              disabled={!auth?.token}
-              onClick={() => setOpenWishlistsWindow(true)}
-            >
-              <FavoriteBorderIcon />
-            </IconButton>
-          </>
         )}
-      </CardActions>
-    </Card>
+        <Link to={`/hotels/hotel/${hotelData.id}`}>
+          <CardActionArea onClick={() => updateHotel({ ...hotelData })}>
+            <CardMedia
+              component="img"
+              height="140"
+              image={hotelData.image || links.noImage}
+              alt={`Hotel ${hotelData.index + 1}`}
+            />
+            <CardContent sx={{ height: "20vh" }}>
+              <Typography gutterBottom variant="h6" component="div">
+                {hotelData.title}
+              </Typography>
+              <Typography variant="body1" color="text.secondary">
+                {truncText(hotelData.description, 70)}
+              </Typography>
+            </CardContent>
+          </CardActionArea>
+        </Link>
+        <CardActions>
+          {isSaved ? (
+            <>
+              <Typography variant="body1" color="text.secondary">
+                <b>
+                  {currentCurrency?.sign}
+                  {Math.round(
+                    +(
+                      hotelData?.price *
+                      currencyList?.rates[currentCurrency?.name]
+                    ).toFixed(1)
+                  )}{" "}
+                  / day
+                </b>
+              </Typography>
+              <IconButton
+                color="secondary"
+                sx={hotelCardStyles.rightAlign}
+                disabled={!Object.keys(currentUser)?.length}
+                onClick={() =>
+                  unsaveHotelFromWishlist({
+                    hotelData,
+                    isSaved,
+                    currentUser,
+                    userUpdate,
+                    userUpdateError,
+                    sendSnackbar,
+                  })
+                }
+              >
+                <FavoriteIcon />
+              </IconButton>
+            </>
+          ) : (
+            <>
+              <Typography variant="body1" color="text.secondary">
+                <b>
+                  {currentCurrency?.sign || "$"}
+                  {Math.round(
+                    +(
+                      hotelData?.price *
+                      currencyList?.rates[currentCurrency?.name]
+                    ).toFixed(1) || hotelData.price
+                  )}{" "}
+                  / day
+                </b>
+              </Typography>
+              <IconButton
+                color="secondary"
+                sx={hotelCardStyles.rightAlign}
+                disabled={!Object.keys(currentUser)?.length}
+                onClick={() => setOpenWishlistsWindow(true)}
+              >
+                <FavoriteBorderIcon />
+              </IconButton>
+            </>
+          )}
+        </CardActions>
+      </Card>
+    </Preloader>
   );
 };
