@@ -5,13 +5,10 @@ import Box from "@mui/material/Box";
 import { ChangePassword } from "./../components/Profile/ChangePassword";
 import { ChangeAvatar } from "./../components/Profile/ChangeAvatar";
 import { ChangePersonalData } from "./../components/Profile/ChangePersonalData";
-import { Preloader } from "./../components/Auxiliary/Preloader";
-import { RootState } from "../helpers/types";
 import { pagesStyles } from "./pagesStyles";
 import { TabPanelProps } from "../server/api/api-models";
-import { useSelector } from "react-redux";
 
-function TabPanel(props: TabPanelProps) {
+const TabPanel = (props: TabPanelProps) => {
   const { children, value, index, ...other } = props;
   return (
     <div
@@ -24,18 +21,16 @@ function TabPanel(props: TabPanelProps) {
       {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
     </div>
   );
-}
+};
 
-function tabsProps(index: number) {
+const tabsProps = (index: number) => {
   return {
     id: `vertical-tab-${index}`,
     "aria-controls": `vertical-tabpanel-${index}`,
   };
-}
+};
 
 export const Profile = () => {
-  const promise = useSelector((state: RootState) => state.promise);
-
   const [value, setValue] = useState(0);
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
@@ -58,25 +53,13 @@ export const Profile = () => {
       </Tabs>
       <div style={pagesStyles.profile.tabPanel}>
         <TabPanel value={value} index={0}>
-          <Preloader
-            promiseName={"uploadAvatar"}
-            promiseState={promise}
-            sub={<ChangeAvatar />}
-          />
+          <ChangeAvatar />
         </TabPanel>
         <TabPanel value={value} index={1}>
-          <Preloader
-            promiseName={"signin"}
-            promiseState={promise}
-            sub={<ChangePersonalData />}
-          />
+          <ChangePersonalData />
         </TabPanel>
         <TabPanel value={value} index={2}>
-          <Preloader
-            promiseName={"signin"}
-            promiseState={promise}
-            sub={<ChangePassword />}
-          />
+          <ChangePassword />
         </TabPanel>
       </div>
     </Box>
